Axios Api running
I was able to do a lot of work throughout the week on this. Forgot to push the code on a daily basis. This code has the Home Page set up with many formulas. Decided to not use the search results page in order to simplify this code. Able to add a next page and previous page button as well. Working on being able to select a movie card to pull up more information.