big update.  read description
finished the contact form using email.js. able to add a success and fail message.

also added a loading state on the search results.

redid the movie info page to be a little cleaner.  Alot of css changes on this  push.

Will be working on making this dynamic for phones and tablets.